backend/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-wordpress-xmlrpc==2.3
mailchimp3==3.0.18
SQLAlchemy==2.0.25
python-dotenv==1.0.0
schedule==1.2.1
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3